app/requirements.txt
django==3.2.*
python-decouple==3.4.*
djangorestframework==3.12.*
gunicorn==20.1.*
psycopg2-binary==2.9.1
pytest-django==4.1.0
pytest-factoryboy==2.0.3
